Use new core

+++ scripts/install_hooks_linux.sh
@@ -6,7 +6,6 @@ if [ ! -d ".git" ]; then
 fi
 
 mkdir -p .git/hooks
-mkdir -p .git/hooks/pre-commit
-mv .git/hooks/pre-commit .git/hooks/pre-commit.bak
+[ -f .git/hooks/pre-commit ] && mv .git/hooks/pre-commit .git/hooks/pre-commit.bak
 cp scripts/hooks/pre-commit.linux .git/hooks/pre-commit
 chmod +x .git/hooks/pre-commit
